Test data output redone in json document
Updated the program to use a comprehensive JSON "document" to store the
results of any individual test, with the settings file embedded. The
intention was to put in all information necessary to recreate the test,
with the excpetion of the slight potential timing variations in the game
engine execution.  The JSON output was test-loaded into Python to verify
the validity of the encoding.