Start frame counter
Added StartFrameCounter to the start method 1 so that the program counts
several frames forward before beginning the process of aiming to allow for
the imprecision in the voice at the onset of volume.

Fixed a bug where the program could continue to run trials while the
interrupt box was active if the engine was using StartMethod 1 and the
subject was still voicing.  Fixed the bug by adding a Pause state to the
game engine.